adaptive-crypto-agility/requirements.txt
# Core cryptographic libraries
cryptography>=41.0.0

# Post-quantum cryptography (Open Quantum Safe)
liboqs-python>=0.10.0

# Networking & utilities
requests>=2.31.0
colorama>=0.4.6
tabulate>=0.9.0
